working on test suite checkpoint to split
