+++ test-suite/Makefile.am
@@ -2,18 +2,20 @@ AM_CPPFLAGS = -Wall -Werror -I$(top_srcdir)/src $(HWLOC_CFLAGS) $(TAP_CFLAGS)
 AM_LDFLAGS = -rpath $(TAP_LIBDIR) 
 LDADD = $(top_srcdir)/src/libmpibind.la $(TAP_LIBS) $(HWLOC_LIBS)
 
-TESTS = \
-    error.t \
-    environment.t \
-    coral_lassen.t \
-    epyc_corona.t \
-    coral_ea.t \
-    cts1_quartz.t
+TEST_EXTENSIONS = .t .py
 
-TEST_EXTENSIONS = .t
+# C Tests
 
 T_LOG_DRIVER = env AM_TAP_AWK='$(AWK)' $(SHELL) \
-            $(top_srcdir)/config/tap-driver.sh
+			$(top_srcdir)/config/tap-driver.sh
+
+PROGS = \
+	error.t \
+	environment.t \
+	coral_lassen.t \
+	epyc_corona.t \
+	coral_ea.t \
+	cts1_quartz.t
 
 coral_lassen_t_SOURCES = coral-lassen.c test_utils.c test_utils.h
 epyc_corona_t_SOURCES = epyc-corona.c test_utils.c test_utils.h
@@ -23,25 +25,25 @@ error_t_SOURCES = error.c test_utils.c test_utils.h
 environment_t_SOURCES = environment.c test_utils.c test_utils.h
 
 if HAVE_LIBTAP
-check_PROGRAMS = $(TESTS)
+check_PROGRAMS = $(PROGS)
 endif
-TEST_EXTENSIONS = .t .py
+
+# Python Tests
 PY_LOG_DRIVER = $(PYTHON) $(top_srcdir)/config/tap-driver.py
 
 PYSCRIPTS = \
 	python/wrapper.py
 
-TESTS = \
-	$(PYSCRIPTS)
-	
-#if HAVE_LIBTAP
-#check_PROGRAMS = $(TESTS)
-#endif
-
 AM_TESTS_ENVIRONMENT = \
 	export PYTHONPATH="$(abs_top_builddir)/bindings/python:$(abs_top_builddir)/test-suite/python::$$PYTHONPATH";\
 	export PYTHON="${PYTHON}";\
-	cp $(top_srcdir)/config/tap-driver.py;
+	cp $(top_srcdir)/test-suite/tap-driver.py $(top_srcdir)/config/tap-driver.py;
+
+# Run both Python and C Tests
+TESTS = \
+	$(PROGS)
+
+# $(PYSCRIPTS)
 
 clean-local:
 	rm -rf trash-directory.* test-results .prove *.broker.log */*.broker.log *.output python/__pycache__
